Move dns package to internal
